+++ api/accounts/models.py
@@ -1,9 +1,8 @@
 from django.db import models
 from django.contrib.auth.models import User
 from django.apps import apps
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-# Topic = apps.get_model('cotrip_app', 'Topic')
+from ..cotrip_app.models import Topic
+
 
 
 class Profile(models.Model):
@@ -24,12 +23,3 @@ class Profile(models.Model):
         return self.first_name + " " + self.last_name
 
 
-@receiver(post_save, sender=User)
-def create_user_profile(sender, instance, created, **kwargs):
-    if created:
-        Profile.objects.create(user=instance)
-
-
-@receiver(post_save, sender=User)
-def save_user_profile(sender, instance, **kwargs):
-    instance.profile.save()
